fix: Store `__elements__` as non-enumerable property on definitions
The `__elements__` property was set as a normal, enumerable property
on CSN definitions.  If that CSN is put in some compiler backends,
the compiler may handle it as a normal CSN property, which could fail
further down.  This is only an issue if the CSN is put directly into
a backend (e.g. the EDMX renderer) and not compiled beforehand.

The compiler may change its behavior to ignore unknown properties (or
those that don't match "our" properties such as `__abc__`) in the
future, but that needs to be discussed first.