Update tests for the New OData Protocol Adapter (Noah)
When serving the service document with the feature toggle `process.env.CDS_FEATURES_ODATA__NEW__ADAPTER = "true"`
enabled, the ordering of collections in the XML format changes slightly. However, ordering collections is not essential
and can vary between implementations.

This commit updates the tests to include separate snapshots for the Legacy OData Protocol Adapter (OKRA) and the new
Noah, as the observed ordering does not follow any specific pattern.